Refactor some logic from pop to program [WIP]

+++ src/evoasm-pop-params.h
@@ -23,8 +23,8 @@ typedef struct evoasm_pop_params_s {
   evoasm_domain_t *domains[EVOASM_POP_PARAMS_MAX_PARAMS];
   uint8_t n_params;
   uint32_t deme_sizes[EVOASM_POP_MAX_DEPTH + 1];
-  uint32_t max_program_sizes[EVOASM_POP_MAX_DEPTH];
-  uint32_t min_program_sizes[EVOASM_POP_MAX_DEPTH];
+  uint16_t max_program_sizes[EVOASM_POP_MAX_DEPTH];
+  uint16_t min_program_sizes[EVOASM_POP_MAX_DEPTH];
   uint8_t depth;
   uint32_t mut_rate;
   evoasm_prng_state_t seed;
